requirements.txt
easydict==1.9
numpy==1.22.3
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.8.0
torch==1.11.0+cu113
tqdm==4.64.0
